Add files via upload
the objective of this project was to build a model to accurately detect the presence of Parkinson’s disease in an individual.
In this Python machine learning project, using the Python libraries scikit-learn, numpy, pandas, and xgboost, i l build a model using an XGBClassifier. i then load the data, get the features and labels, scale the features, then split the dataset, build an XGBClassifier, and then calculate the accuracy of the model.